refactor: split bindings into separate files
